requirements.txt
asgiref==3.4.1
Django==4.0.1
nulltype==2.3.1
plaid-python==8.9.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
